move format and result in separate files and reexport
