+++ src/lib.rs
@@ -17,23 +17,25 @@
 
 use std::{convert::TryFrom, fmt::Debug};
 
+pub use format::{PgFormat, UnrecognizedFormat};
 pub use hand_shake::{Process as HandShakeProcess, Request as HandShakeRequest, Status as HandShakeStatus};
 pub use message_decoder::{MessageDecoder, Status as MessageDecoderStatus};
 pub use messages::{BackendMessage, ColumnMetadata, FrontendMessage};
+pub use result::{Error, Result};
 
 mod cursor;
+mod format;
 mod hand_shake;
 mod message_decoder;
 /// Module contains backend messages that could be send by server implementation
 /// to a client
 mod messages;
 mod request_codes;
+mod result;
 mod types;
 
 /// Connection key-value params
 pub type ClientParams = Vec<(String, String)>;
-/// Protocol operation result
-pub type Result<T> = std::result::Result<T, Error>;
 
 /// PostgreSQL OID [Object Identifier](https://www.postgresql.org/docs/current/datatype-oid.html)
 pub type Oid = u32;
@@ -41,54 +43,3 @@ pub type Oid = u32;
 pub(crate) type ConnId = i32;
 /// Connection secret key
 pub(crate) type ConnSecretKey = i32;
-
-/// PostgreSQL formats for transferring data
-/// `0` - textual representation
-/// `1` - binary representation
-#[derive(Debug, PartialEq, Clone, Copy)]
-pub enum PgFormat {
-    /// data from/to client should be sent in text format
-    Text,
-    /// data from/to client should be sent in binary format
-    Binary,
-}
-
-impl TryFrom<i16> for PgFormat {
-    type Error = UnrecognizedFormat;
-
-    fn try_from(value: i16) -> std::result::Result<Self, Self::Error> {
-        match value {
-            0 => Ok(PgFormat::Text),
-            1 => Ok(PgFormat::Binary),
-            other => Err(UnrecognizedFormat(other)),
-        }
-    }
-}
-
-/// Represents an error if frontend sent unrecognizable format
-/// contains the integer code that was sent
-#[derive(Debug)]
-pub struct UnrecognizedFormat(i16);
-
-/// `Error` type in protocol `Result`. Indicates that something went not well
-#[derive(Debug, PartialEq)]
-pub enum Error {
-    /// Indicates that the current count of active connections is full
-    ConnectionIdExhausted,
-    /// Indicates that incoming data is invalid
-    InvalidInput(String),
-    /// Indicates that incoming data can't be parsed as UTF-8 string
-    InvalidUtfString,
-    /// Indicates that incoming string is not terminated by zero byte
-    ZeroByteNotFound,
-    /// Indicates that frontend message is not supported
-    UnsupportedFrontendMessage,
-    /// Indicates that protocol version is not supported
-    UnsupportedVersion,
-    /// Indicates that client request is not supported
-    UnsupportedRequest,
-    /// Indicates that during handshake client sent unrecognized protocol version
-    UnrecognizedVersion,
-    /// Indicates that connection verification is failed
-    VerificationFailed,
-}
